+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## unreleased
-* Fix date offset in dashboard widget in WP 5.3+ environments with mixed timezones
+* Fix date offset in dashboard widget in WP 5.3+ environments with mixed timezones (#167)
+* Allow to deactivate the nonce check during JavaScript tracking (#168)
 
 ## 1.7.2
 * Prevent JavaScript tracking from raising 400 for logged-in users, if tracking is disabled (#159)